Move HibernateConfiguration into net.corda.node.services.persistence
