+++ node/src/test/kotlin/net/corda/node/services/persistence/DBTransactionStorageTests.kt
@@ -13,7 +13,6 @@ import net.corda.core.transactions.WireTransaction
 import net.corda.finance.schemas.CashSchemaV1
 import net.corda.finance.schemas.SampleCashSchemaV2
 import net.corda.finance.schemas.SampleCashSchemaV3
-import net.corda.node.services.database.HibernateConfiguration
 import net.corda.node.services.schema.HibernateObserver
 import net.corda.node.services.schema.NodeSchemaService
 import net.corda.node.services.transactions.PersistentUniquenessProvider
